+++ src/main/kotlin/com/meli/domain/pessoa/usecase/PessoaAllGetCpf.kt
@@ -2,5 +2,5 @@ package com.meli.domain.pessoa.usecase
 
 
 interface PessoaAllGetCpf {
-    fun process(cpfPessoa: Int): Any?
+    fun process(cpfPessoa: Long): Any?
 }
