+++ src/main/kotlin/com/meli/domain/pessoa/usecase/PessoaAllSave.kt
@@ -2,5 +2,11 @@ package com.meli.domain.pessoa.usecase
 
 
 interface PessoaAllSave {
-    fun process(idPessoa: Long, nomePessoa: String, sobrenomePessoa: String, carroPessoa: String?, cpfPessoa: Int)
+    fun process(
+        idPessoa: Long,
+        nomePessoa: String,
+        sobrenomePessoa: String,
+        carroPessoa: String?,
+        cpfPessoa: Long
+    )
 }
